Reworked types.h to include types from other subsystems
