+++ micromouse_main/src/devices/encoders.cpp
@@ -2,6 +2,7 @@
 
 
 #include <Arduino.h>
+
 #include "encoders.h"
