+++ micromouse_main/src/devices/encoders.h
@@ -10,4 +10,5 @@ void encoderSetup(unsigned char id, unsigned char pinA, unsigned char pinB);
 // Reads from an encoder given an id
 int readEncoder(unsigned char id);
 
+
 #endif //_ENCODERS_H_
