+++ micromouse_main/src/devices/sensors.h
@@ -23,7 +23,7 @@
 
 
 #include "Adafruit_VL6180X.h"
-#include "../settings.h"
+
 #include "../types.h"
